moving daemon to api, wip

+++ public-api/src/database/task.rs
@@ -6,7 +6,8 @@ use url::Url;
 use uuid::Uuid;
 
 use crate::error::Result;
-const TASK_MAX_RETRIES: i16 = 5;
+
+const TASK_MAX_RETRIES: u8 = 5;
 
 #[derive(Deserialize, Serialize, Debug, Copy, Clone, sqlx::Type)]
 #[sqlx(type_name = "task_status")]
@@ -27,13 +28,13 @@ pub struct Task {
     pub created_at: DateTime<Utc>,
     pub updated_at: DateTime<Utc>,
     pub next_delivery: DateTime<Utc>,
-    pub retries: Option<i16>,
+    pub retries: Option<i8>,
     pub fail_reason: Option<String>,
 }
 
 impl Task {
     pub fn should_retry(&self) -> bool {
-        self.retries.unwrap_or(0) < TASK_MAX_RETRIES
+        self.retries.unwrap_or(0) < TASK_MAX_RETRIES as i8
     }
 }
 
@@ -99,7 +100,7 @@ pub async fn update(
     pool: &Pool<Postgres>,
     task: &Task,
     status: TaskStatus,
-    retries: Option<i16>,
+    retries: Option<i8>,
     fail_reason: Option<String>,
 ) -> Result<()> {
     let sql =
